requirements.txt
amqp==5.0.6
asgiref==3.3.1
attrs==21.2.0
billiard==3.6.4.0
black==21.9b0
celery==5.1.2
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.1.7
djangorestframework==3.12.2
iniconfig==1.1.1
kombu==5.1.0
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.9.0
platformdirs==2.3.0
pluggy==1.0.0
prompt-toolkit==3.0.21
py==1.10.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
regex==2021.8.28
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
vine==5.0.0
wcwidth==0.2.5
RPi.GPIO==0.7.0